backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]
supabase-py
pymongo
python-dotenv